lyft-forage-master/system/engine.py
from .serviceable import Serviceable
from abc import ABC, abstractmethod

class Engine(Serviceable, ABC):
    @abstractmethod
    def needs_service(self) -> bool:
        pass

class CapuletEngine(Engine):
    SERVICE_INTERVAL = 30000

    def __init__(self, last_service_mileage, current_mileage):
        self.last_service_mileage = last_service_mileage
        self.current_mileage = current_mileage

    def needs_service(self) -> bool:
        return self.current_mileage - self.last_service_mileage >= self.SERVICE_INTERVAL

class SternmanEngine(Engine):
    def __init__(self, warning_light_on):
        self.warning_light_on = warning_light_on

    def needs_service(self) -> bool:
        return self.warning_light_on

class WilloughbyEngine(Engine):
    SERVICE_INTERVAL = 60000

    def __init__(self, last_service_mileage, current_mileage):
        self.last_service_mileage = last_service_mileage
        self.current_mileage = current_mileage

    def needs_service(self) -> bool:
        return self.current_mileage - self.last_service_mileage >= self.SERVICE_INTERVAL
